refactor: split ast-related code
