+++ src/ast/clone.ts
