+++ src/ast/hash.ts
